+++ tests/test_works.py
@@ -10,13 +10,16 @@ def test_works(scenario: Scenario) -> None:
             """
             from extended_mypy_django_plugin import Concrete, DefaultQuerySet
 
+            from typing import TypeVar
             from myapp.models import Parent, Child1, Child2
 
-            T_Child = Concrete.type_var("T_Child", Parent)
+            T_Child = TypeVar("T_Child", bound=Concrete[Parent])
 
 
             def make_child(child: type[T_Child]) -> T_Child:
-                return child.objects.create()
+                created = child.objects.create()
+                assert isinstance(created, child)
+                return created
 
 
             def make_any_queryset(child: type[Concrete[Parent]]) -> DefaultQuerySet[Parent]: